requirements.txt
# --------- pytorch --------- #
torch==2.8.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
neptune==1.14.0
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

# --------- data science --------- #
biopython==1.85
matplotlib==3.10.7
faiss-cpu==1.12.0
numba==0.62.1
parasail==1.3.4
blosum==2.0.3
tqdm==4.67.1
pillow==11.3.0